[CLEANUP-3] Complete Issue #153 Security Simplification - Final commit
